requirements.txt
tensorflow
numpy
pillow
opencv-python
matplotlib
pandas
tqdm
keras-tuner
scikit-learn
seaborn